Move ArrowDrawer to graphics package
