+++ src/main/java/br/com/etyllica/awt/AWTArrowDrawer.java
@@ -3,7 +3,7 @@ package br.com.etyllica.awt;
 import java.awt.BasicStroke;
 import java.awt.Color;
 
-import br.com.etyllica.core.drawer.ArrowDrawer;
+import br.com.etyllica.core.graphics.ArrowDrawer;
 import br.com.etyllica.core.graphics.Graphic;
 import br.com.etyllica.core.input.mouse.MouseState;
 import br.com.etyllica.theme.mouse.ArrowTheme;
