+++ src/main/java/br/com/etyllica/core/InnerCore.java
@@ -13,11 +13,11 @@ import br.com.etyllica.core.animation.script.SingleIntervalAnimation;
 import br.com.etyllica.core.context.Application;
 import br.com.etyllica.core.context.Context;
 import br.com.etyllica.core.context.UpdateIntervalListener;
-import br.com.etyllica.core.drawer.ArrowDrawer;
 import br.com.etyllica.core.event.GUIEvent;
 import br.com.etyllica.core.event.KeyEvent;
 import br.com.etyllica.core.event.PointerEvent;
 import br.com.etyllica.core.event.PointerState;
+import br.com.etyllica.core.graphics.ArrowDrawer;
 import br.com.etyllica.core.graphics.Graphic;
 import br.com.etyllica.core.graphics.Monitor;
 import br.com.etyllica.core.i18n.Language;
